lara/cheat: port Lara_Cheat_Teleport
