+++ src/libtrx/include/libtrx/game/lara/cheat.h
@@ -11,5 +11,4 @@ bool Lara_Cheat_KillEnemy(int16_t item_num);
 bool Lara_Cheat_OpenNearestDoor(void);
 bool Lara_Cheat_EnterFlyMode(void);
 bool Lara_Cheat_ExitFlyMode(void);
-extern bool Lara_Cheat_Teleport(
-    int32_t x, int32_t y, int32_t z, int16_t room_num);
+bool Lara_Cheat_Teleport(int32_t x, int32_t y, int32_t z, int16_t room_num);
