+++ src/tr1/meson.build
@@ -141,7 +141,6 @@ sources = [
   'game/item_actions/sand.c',
   'game/item_actions/stairs2slope.c',
   'game/item_actions/turn_180.c',
-  'game/lara/cheat.c',
   'game/lara/col.c',
   'game/lara/common.c',
   'game/lara/control.c',
